SingularModes.jl

Software package written in Julia for the propagation of transverse laser modes through turbulent media.

Installation

Prerequisites: Install Julia

You need a running copy of Julia to use this package. We recommend to install it via the package manager of your system, e.g., pacman or apt-get on Arch or Debian-based Linux, respectively.

sudo pacman -S julia

# Debian-based Linux
sudo apt-get install julia

On macOS, Julia can be installed via the package manager Homebrew which is recommended for easy updates later on. Install Homebrew and Julia:

# macOS
# Install Homebrew
/bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
# Install Julia
brew install julia

Install the SingularModes.jl package

Once Julia is running on your system, the SingularModes package can straightforwardly be installed via Julia's internal package manager.

julia> ]
pkg> add SingularModes

Package structure

This software package is structured as follows:

.
│
├── src                    [Julia modules]
│   ├── BasicFunctions.jl  [library of general functions and definitions]
│   ├── FigureDefaults.jl  [default settings for plotting]
│   ├── ModeGeneration.jl  [library for transverse mode generation]
│   ├── Phasescreens.jl    [library for phase screen generation]
│   ├── Plotting.jl        [library for plotting routines]
│   ├── PowerSpectra.jl    [library of power spectra for different media]
│   ├── Propagation.jl     [library for all functions related to propagation]
│   └── Statistics.jl      [library for the statistical analysis of phase screens]
│
├── docs                   [documentation of the package]
│   └── ...
│
├── examples               [code examples]
│   ├── static             [cached arrays of LG/ZK/KL modes for faster computation]
│   │
│   ├── parameters.jl      [sample parameter file]
│   ├── LG-Beams.jl        [sample code for the propagtion of Lagurre-Gaussian beams]
│   ├── Singularmodes.jl   [sample code for the computation of singular modes]
│   └── ...
│
├── tests                  [code tests]
│   └── ...
│
├── LICENSE                [GPL-3.0-or-later]
├── README.md              [Short documentation]
│
└── ...

The constituent Julia modules are documented below.

BasicFunctions.jl

SingularModes.BasicFunctions.cart2pol — Method

cart2pol(x, y)

Convert Cartesian coordinates to polar coordinates.

SingularModes.BasicFunctions.ft — Method

ft(g, delta=1)

Multi dimensional Fourier transfrom with optional gridspacing d = 1`.

SingularModes.BasicFunctions.ift — Method

ift(g, delta_f=1)

Inverse multi dimensional Fourier transfrom. See also ft.

SingularModes.BasicFunctions.inner_circ — Method

inner_circ(a; nans=false)

Set all elements outside of an inscribed circle to zero or NaN (nans=true).

SingularModes.BasicFunctions.meshgrid — Method

meshgrid(x, y)

Rectangular meshgrid array from the vectors x and y.

SingularModes.BasicFunctions.msum — Method

msum(mode)

Total power of a mode.

SingularModes.BasicFunctions.norm_coeff — Method

norm_coeff(vec)

Normalizes the vector vec s.t. sum(abs.2(vec))=1.

SingularModes.BasicFunctions.norm_mode — Method

norm_mode(mode)

Normalize the intensity of a mode such that the total power yields one.

SingularModes.BasicFunctions.pol2cart — Method

pol2cart(th, r)

Convert polar to Cartesian coordinates.

SingularModes.BasicFunctions.print_paras — Method

print_paras(datdir,
    Rphys, outin, zzR,
    N, Nin,
    Rin, RoutC, Rout, din, dout, wvl, z,
    w0, nmax, jlmax, zR, nhigh, ratio,
    spec, r0, l0, L0, sh, jzmax,
    rs, ns, sr2,
    vmean, vstd, tv, deltat, Ts)

Print parameters from parameters.jl.

SingularModes.BasicFunctions.rayleigh_waist — Method

rayleigh_waist(z, wvl)

Determine beam waist such that z = zR (Rayleigh length). This leads to the smallest possible output aperture.

SingularModes.BasicFunctions.rect — Method

rect(x; w=1, nans=false)

Rectangle function returns one if x is within (-w/2, w/2) else zero or NaN (nans=true).

SingularModes.BasicFunctions.struct_fcn2_ft — Method

struct_fcn2_ft(ph, mask, delta)

Structure function based on FT from Schmidt (2010). Very sensitive to boundaries, needs zeropadding. Not tested. Recommended not to use.

SingularModes.BasicFunctions.wopt — Method

wopt(D, Df)

Theoretical beam width for optimal vacuum propagation.

SingularModes.BasicFunctions.zeropad — Method

zeropad(a, s)

Zeropad a matrix to the desired side length s.

FigureDefaults.jl

Plots.default — Function

default()

Set defaults for plotting figures with the Plots package.

ModeGeneration.jl

SingularModes.ModeGeneration._default_N — Function

_default_N(N, fpointsname)

Default gridsize N for LG and ZK modes. See also lg, zk, points2grid.

SingularModes.ModeGeneration.bcl — Method

bcl(uin)

Fourier-based beam correlation length of an input beam profile uin`.

SingularModes.ModeGeneration.besselbeam — Method

besselbeam(x, y; z=0, wvl=1550e-9, beta=0, l=0, paraxial=true)

Pointwise Bessel mode function. For non-paraxial wave equation set paraxial=false.

SingularModes.ModeGeneration.besselgaussian — Method

besselgaussian(x, y; w0=1, z=0, wvl=1550e-9, beta=0, l=0)

Pointwise Bessel-Gaussian mode function. See also besselbeam.

SingularModes.ModeGeneration.bg_mode — Method

bg_mode(; N=256, R=1, kwargs...)

Shorthand for a single BG mode on a grid. See also besselgaussian, points2grid.

SingularModes.ModeGeneration.hg — Method

hg(x, y; w0x=1, w0y=1, w0=1, z=0, wvl=1550e-9, m=0, n=0, d=1)

Pointwise Hermite-Gaussian mode function.

SingularModes.ModeGeneration.hg_1d — Method

hg_1d(x, y; w0=1, z=0, wvl=1550e-9, m=0, d=1)

Pointwise one-dimensional Hermite-Gaussian mode function.

SingularModes.ModeGeneration.hg_mode — Method

hg_mode(; N=256, R=1, kwargs...)

Shorthand for a single HG mode on a grid. See also hg, points2grid.

SingularModes.ModeGeneration.j2mn — Method

j2mn(j) = j2mn(j, hg=true)

Shorthand for reconversion of HG Noll indices. See also j2pl.

SingularModes.ModeGeneration.j2nm — Method

j2nm(j) = j2pl(j, zk=true)

Shorthand for reconversion of ZK Noll indices. See also j2pl.

SingularModes.ModeGeneration.j2nmax — Method

j2nmax(j)

Maximal radial order nmax for a given superindex j. See also j2pl.

SingularModes.ModeGeneration.j2pl — Method

j2pl(j, zk=false)

Reconversion of Noll superindex j to individual indices p, l for LG modes and n, m for ZK modes (with zk=true).

SingularModes.ModeGeneration.laguerre — Method

laguerre(n, alpha, x)

Associated/generalized Laguerre polynomial with radial order n and azimuthal index alpha. Definition from wikipedia.

SingularModes.ModeGeneration.lg — Method

lg(x, y; w0=1, z=0, wvl=1550e-9, p=0, l=0, d=1, j=0)

Pointwise Lagurre-Gaussian mode function. See also laguerre.

SingularModes.ModeGeneration.lg_mode — Method

lg_mode(; N=256, R=1, kwargs...)

Shorthand for a single LG mode on a grid. See also lg, points2grid.

SingularModes.ModeGeneration.make_grid — Function

make_grid(N, R)

Meshgrid of size N x N with sidelength 2R.

SingularModes.ModeGeneration.mn2j — Method

mn2j(m, n) = pl2j(m, n, hg=true)

Shorthand for HG Noll indices. See also pl2j.

SingularModes.ModeGeneration.mode_pyramid — Method

mode_pyramid(fpoints, nmax, N, R; jmax=0, threads=true, kwargs...)

Generate a pyramid array of a given mode via fpoints, i.e., kwargs... -> fpoints. See also lg, zk.

SingularModes.ModeGeneration.mode_pyramid_load — Method

mode_pyramid_load(fpoints, nmax, N, R; kwargs...)

Load the mode pyramid from file (currently disabled). See also mode_pyramid.

SingularModes.ModeGeneration.nm2j — Method

nm2j(n, m) = pl2j(n, m, zk=true)

Shorthand for ZK Noll indices. See also pl2j.

SingularModes.ModeGeneration.nmax2j — Method

nmax2j(j)

Maximal superindex j for a given radial order nmax. See also j2pl.

SingularModes.ModeGeneration.noll_pyramid — Method

noll_pyramid(nmax)

Print a pyramid of Noll indices up to the radial order nmax. See also j2pl, pl2j.

SingularModes.ModeGeneration.pcl — Method

pcl(l; w0=1)

Phase correlation length based on Eq.~(11) in [Leonhard (2015)].

SingularModes.ModeGeneration.pcl_arc — Method

pcl_arc(l; w0=1)

Phase correlation arc.

SingularModes.ModeGeneration.pcl_bg — Method

pcl_bg(l, beta; w0=1, mod=false)

Phase correlation length for Bessel-Gaussian beams, as used in [Bachmann (2024)].

SingularModes.ModeGeneration.pcl_geo — Method

pcl_geo(l; w0=1)

Geometric phase correlation length.

SingularModes.ModeGeneration.pcl_lim — Method

pcl_lim(l; w0=1)

Large l limit of the phase correlation length.

SingularModes.ModeGeneration.pcl_mod — Method

pcl_mod(l; w0=1)

Modified phase correlation length based on an isosceles triangle.

SingularModes.ModeGeneration.pl2j — Method

pl2j(p, l; zk=false)

Double to single index conversion on basis of Noll's convention. Use p, l for LG modes, or n, m for ZK modes (with zk=true) or m, n for HG modes (with hg=true).

SingularModes.ModeGeneration.points2grid — Method

points2grid(fpoints; N=1024, R=1, kwargs...)

Evaluate the function fpoints on an N x N grid of sidelength 2R (default N=1024). Keywords arguments kwargs... are passed to fpoints. See also _default_N.

SingularModes.ModeGeneration.r_z — Method

r_z(r, n, m)

Radial function of the Zernike polynomials based on [Noll (1976)], Eq. (1).

SingularModes.ModeGeneration.rho_mean_bg — Method

rhomeanbg(l, beta; w0=1)

Mean radius of a Bessel-Gaussian beam.

SingularModes.ModeGeneration.zk — Method

zk(x, y; R=1, j=1, angle=0)

Zernike polynomial with Noll index j, radius R and rotation angle angle. See also r_z.

SingularModes.ModeGeneration.zk_mode — Method

zk_mode(; N=1024, R=1, kwargs...)

Shorthand for a single ZK mode on a grid. See also zk, points2grid.

SingularModes.ModeGeneration.zka_load — Method

zka_load(nmax, darray; N=1024)

Generate an array of ZK pyramids for the multi-step propagation. Load from file if it exsists.

PhaseScreens.jl

SingularModes.PhaseScreens.fourier_shift — Method

fourier_shift(
    a, dshift;
    d=1
    )

Subpixel Fourier shift of a two-dimensional array a by distance dshift.

SingularModes.PhaseScreens.get_phzT — Method

get_phzT(
    phzB, Rarray, Tarray, vmean, vstd;
    Norg=1024, rsw=0, kwargs...
    )

Returns a time array of z-arrays of wind-shifted phase screens based on a large underlaying screen z-array phzB with mean transverse wind velocity vmean and its standard deivation vstd at time time. The random seed for the wind velocity distribution defaults to rsw=0, and the subscreen side length to Norg=1024 pixels.

SingularModes.PhaseScreens.pFq_reg — Method

pFq(alist, blist, z)

Regularized hypergeometric function. Relies on HypergeometricFunctions.jl.

SingularModes.PhaseScreens.ph_ft — Method

ph_ft(
    pow_spec, N, R, rs;
    silent=false, kwargs...
    )

Fourier phase screen for a given pow_spec on an N x N array with physical radius R and radom seed rs.

SingularModes.PhaseScreens.ph_hd — Method

ph_hd(
    pow_spec, N, R, rs;
    nmax=9, jzmax=0, zks=zks, phft=phft, nans=false, kwargs...
    )

Hybrid phase screen for a given pow_spec on an N x N array with physical radius R and radom seed rs. Defaults to nmax=9 pyramidal levels for the constituent Zernike screen.

SingularModes.PhaseScreens.ph_kl — Method

ph_kl(
    pow_spec, N, R, rs;
    nmax=9, jzmax=0, zks=zks, kls=kls, phft=phft, klsReturn=false,
    kwargs...
    )

Hybrid Karhunen-Loève phase screen for a given pow_spec on an N x N array with physical radius R and radom seed rs. Defaults to nmax=9 pyramidal levels for the constituent Zernike screen.

SingularModes.PhaseScreens.ph_sh — Method

ph_sh(
    pow_spec, N, R, rs;
    sh=5, kwargs...
    )

Subharmonic Fourier phase screen for a given pow_spec on an N x N array with physical radius R and radom seed rs. Defaults to sh=5 subharmonic levels.

SingularModes.PhaseScreens.ph_x — Method

ph_x(
    pow_spec;
    N=1024, R=1, rs=0, kind="hd", kwargs...
    )

Wrapper function for single phase screens. Supported kinds are "ft", "sh", "zk", "hd", and "kl", see keyword arguments of the corresponding functions. Defaults to a hybrid phase screen with N=1024, R=1, and rs=0.

SingularModes.PhaseScreens.ph_zk — Method

ph_zk(
    pow_spec, N, R, rs;
    nmax=9, jzmax=0, zks=zks, ascorr=ascorr, silent=false, kwargs...
    )

Zernike phase screen for a given pow_spec on an N x N array with physical radius R and radom seed rs. Defaults to nmax=9 pyramidal levels.

SingularModes.PhaseScreens.phz_x — Method

ph_x(
    pow_spec, Rarray;
    N=1024, R=1, rs=0, kind="hd", r0tot=0, kwargs...
    )

Wrapper function for phase screen arrays along the z-axis specied via Rarray. Supported kinds are "ft", "sh", "zk", "hd", and "kl", see keyword arguments of the corresponding functions. Defaults to a hybrid phase screens with N=1024, R=1, and rs=0. The total Fried parameter has either to be specified via r0tot or multiplied manually for non-Kolmogorov turbulence.

SingularModes.PhaseScreens.screen_factor — Method

screen_factor(
    vmean, vstd, tmax, R;
    offset=0.1
    )

Computes the required size of a large, time-shifted phase screen based on the mean transverse wind velocity vmean, its standard deivation vstd, the maximally considered time tmax, and the smaller subscreen's radius R.

SingularModes.PhaseScreens.wind_shift — Method

wind_shift(
    phzB, Norg, Rarray, vmean, vstd, time;
    rsw=0, offset=0.1, threads=true, kwargs...
    )

Returns a z-array of wind-shifted phase screens with resolution Norg x Norg based on a large underlaying screen array phzB with mean transverse wind velocity vmean and its standard deivation vstd at time time. The random seed for the wind velocity distribution defaults to rsw=0, the offset=0.1 defaults to a 10% margin from the large screens boundaries.

SingularModes.PhaseScreens.zk_cov_atm — Method

zk_cov_atm(j, jp, R; L0=Inf, alpha=5/3)

Analytical Zernike covariances for the atmospheric phase power spectra.

SingularModes.PhaseScreens.zk_cov_num — Method

zk_cov_num(power_spec, j, jp, R; kwargs...)

Numerical Zernike covariances for arbitrary power_spec. All kwargs... are passed to power_spec.

SingularModes.PhaseScreens.zk_covmat — Method

zk_covmat(
    pow_spec, nmax, R;
    analytics=true, jmax=0, l0=0, alpha=5/3, kwargs...
    )

Zernike covariance matrix obtained analytically for Kolmogorov and von Karman turbulence , else (or if analytics=false) numerically.

SingularModes.PhaseScreens.zk_overlap — Method

zk_overlap(
    ph, zks
    )

Computes the overlap of a phase screen ph with the Zernike polynomials zks.

Plotting.jl

SingularModes.Plotting.clean_mode_plot — Method

clean_mode_plot(mode; kind=2)

Shorthand for clean mode plot. See also mode_plot.

SingularModes.Plotting.colorbar2d_plot — Function

colorbar2d_plot(ratio=5)

2D colorbar for HSV-RGBA images. See also phasemap.

SingularModes.Plotting.mode_plot — Method

mode_plot(
    mode; R=1, kind=0, legend=true,
    boundary=true, circle=true, yflip=false, clean=false
    )

Plot intensity (kind = 0), phase (kind = 1) or combination (kind = 2) of a mode.

SingularModes.Plotting.phasemap — Method

phasemap(mode; cmap=hsv, yflip=false, square=true)

Translate (complex) matrix into HSV-RGBA image.

SingularModes.Plotting.ps_plot — Method

ps_plot(ph; R=1, r0=0, legend=true, boundary=true)

Phase screen plot with optional scale bar for r0 != 0.

SingularModes.Plotting.s_plot — Method

s_plot(svs; m=-1)

Singularvalue plot with optional indicator at index m.

SingularModes.Plotting.tmat_plot — Method

tmat_plot(tmat; grid=true, kind=0, legend=true, yflip=true, square=false)

Transmission matrix plot of different kinds. See also mode_plot.

PowerSpectra.jl

SingularModes.PowerSpectra._oce_delta — Method

_oce_delta(k, eta, C1)

Help function for ocespec. See also [`ocespec`](@ref).

SingularModes.PowerSpectra.atm_spec — Method

atm_spec(k; L0=Inf, l0=0, alpha=5/3)

Atmospheric phase power spectrum. E.g. for

L0=Inf, l0=0, alpha=5/3: Kolmogorov

L0=Inf, l0=0, alpha=2 : quadratic approximation

L0<Inf, l0=0, alpha=5/3: von Karman

L0<Inf, l0>0, alpha=5/3: Tatarskii.

SingularModes.PowerSpectra.atm_struct — Method

atm_struct(r; L0=Inf)

Theoretical structure function of Kolmogorov (L0=Inf) or von Karman (L0 < Inf) turbulence.

SingularModes.PowerSpectra.oce_spec — Method

oce_spec(k; w=-0.5, eta=0.1)

Oceanic phase power spectrum based on Korotkova (2018).

SingularModes.PowerSpectra.ps2struct — Method

ps2struct(pow_spec, r; kwargs...)

Numerical conversion of a given phase power spectrum to its structure function, the kwargs... are passed to pow_spec. See also atm_spec, oce_spec.

Propagation.jl

SingularModes.Propagation.ang_spec_prop — Method

ang_spec_prop(u1, wvl, d1, d2, z)

Fresnel angular spectrum propagation based on [Schmidt (2010)].

SingularModes.Propagation.cn2 — Method

cn2(r0, wvl, z)

Refractive index structure constant C_n^2.

SingularModes.Propagation.cn2sr — Method

cn2sr(sr2, wvl, z)

Refractive index structure constant C_n^2 via Rytov variance sr2.

SingularModes.Propagation.delta_tmat — Method

detla_tmat(tmat1, tmat2, deltat)

Difference quotient of transmission matrices. BETA!

SingularModes.Propagation.fraunhofer_prop — Method

fraunhofer_prop(u1, wvl, d1, z)

Fraunhofer propagation based on Schmidt (2010).

SingularModes.Propagation.gws_q — Method

gws_q(tmat0, tmat1, tmat2, deltat)

Generalized Wigner-Smith Q. BETA!

SingularModes.Propagation.lens_against_ft — Method

lens_against_ft(u1, wvl, d1, f)

Propagation from the pupil plane to the focal plane of a lens for an object placed against (or just before) the lens.

SingularModes.Propagation.min_ns — Function

min_ns(z, wvl, r0, limit=0.5)

Determine the minimal number of propagation steps while complying with the weak scintillaiton condition (default sr2 < limit = 0.5).

SingularModes.Propagation.modeset_propagation — Method

modeset_propagation(modes, wvl, Rarray, zarray, phz)

Propagate a whole set of modes through a channel. See also propagate.

SingularModes.Propagation.multi_step_variable — Method

multi_step_variable(uin, wvl, darray, zarray, phz; track=false)

Multi step propagation along a channel described by darray, zarray and phz.

SingularModes.Propagation.principal_modes — Method

principal_modes(gws, lgs)

Principal modes. BETA!

SingularModes.Propagation.propagate — Method

propagate(mode, wvl, Rarray, zarray, phz)

Propagate through a channel (including numerical padding and cropping). See also multi_step_variable.

SingularModes.Propagation.r0step — Method

r0step(r0, ns)

Fried parameter of individual propagation steps (assumes equal step length), based on Eq.~(5.68) in [Sorelli (2019)].

SingularModes.Propagation.rayleigh_length — Method

rayleigh_length(w0, wvl)

Rayleigh length of optical propagation.

SingularModes.Propagation.sigma_R2 — Method

sigma_R2(wvl, z, r0)

Rytov variance.

SingularModes.Propagation.sigma_R2_1 — Method

sigma_R2_1(wvl, z, r0, ns)

Fist (and largest) term of Rytov variance for multi step propagation as given in Eq.~(5.72) in [Sorelli (2019)].

SingularModes.Propagation.svd_lgs_pixel — Method

svd_lgs_pixel(lgs, tmat)

Singular modes via SVD of a transmission matrix tmat with output pixel basis.

SingularModes.Propagation.svd_lgs_square — Method

svd_lgs_square(lgs, lgs_turb, tmat)

Singular modes via SVD of a square transmission matrix tmat.

SingularModes.Propagation.svd_square_wrapper — Method

svd_square_wrapper(Nin, R, w0, nmax, Rarray, zarray, phz)

Wrapper for all steps to obtain the singular modes of a square transmission matrix.

SingularModes.Propagation.transmission_pixel — Method

transmission_pixel(lgs, phz, wvl, Rarray, zarray)

Transmission matrix of a channel with output pixel basis.

SingularModes.Propagation.transmission_square — Method

transmission_square(lgs_vac, lgs_turb)

Square transmission matrix from vacuum propagated (lgs_vac) and turbulence propagated (lgs_turb) modes, i.e. matrix with entries <lgsvac | lgsturb>.

Statistics.jl

SingularModes.Statistics.chunk_avg — Method

chunk_avg(a, c)

Split 1D array into chunks of size c and compute their mean.

SingularModes.Statistics.phase_struct_ft — Method

phase_struct_ft(a, R)

Radial structure function of a 2D array based on Fourier Transform. Needs zeropadding. Not recommended to use, rather use phase_struct_rd.

SingularModes.Statistics.phase_struct_rd — Method

phase_struct_rd(ph, R; rs=0, samples=10^8)

Radial phase structure function of a 2D array based on random sampling.

SingularModes.Statistics.points_in_circ — Method

points_in_circ(radius, samples)

Unifromly distributed random points confined by a circle.

SingularModes.Statistics.radial_avg — Method

radial_avg(a, d)

Average of a 2D array along centered circles.

SingularModes.Statistics.zernike_covariance — Method

zernike_covariance(
    pow_spec;
    nmaxScreen=9,
    nmaxZernike=10,
    samples=10,
    rs=0,
    N=1024,
    R=1,
    zks=zeros(1,1,1),
    kwargs...
    )

Zernike variance of screens for a given power spectrum. See also SingularModes.PhaseScreens.zk_overlap, SingularModes.PhaseScreens.ph_x.